Create Layout Calculator
The first section is the overall menu. This keeps on accepting user input until the user enters "quit", so a while loop is used.The code above is the starting point for our program. It accepts user input, and compares it to the options in the if/elif statements.
The break statement is used to stop the while loop, in case the user inputs "quit".